python-api/requirements.txt
fastapi
pillow
uvicorn[standard]
pydantic
